requirements.txt
requests==2.31.0
playwright==1.40.0
lxml==5.1.0
pyyaml==6.0.1
jinja2==3.1.3
python-dotenv==1.0.0
